@@ -4,7 +4,8 @@ version: 3.1.0
 homepage: https://github.com/TercyoStorck/internationalization
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
+  flutter: ">=2.5.0"
 
 dependencies:
   flutter:
