@@ -9,7 +9,7 @@ export 'package:intl/intl.dart' show DateFormat;
 export 'package:intl/intl.dart' show NumberFormat;
 
 class Internationalization {
-  static BuildContext _context;
+  static late BuildContext _context;
 
   static void of(BuildContext context) => _context = context;
   static BuildContext get context => _context;
